src/main/java/com/huangjie/sell/service/impl/ProductCategoryServiceImpl.java
package com.huangjie.sell.service.impl;

import com.huangjie.sell.dataobject.ProductCategory;
import com.huangjie.sell.repository.ProductCategoryRepository;
import com.huangjie.sell.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Created by huangjie on 2019/7/18.
 */
@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryRepository productCategoryRepository;

    @Override
    public ProductCategory findOne(Integer categoryID) {
        return productCategoryRepository.findOne(categoryID);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryRepository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return productCategoryRepository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryRepository.save(productCategory);
    }
}
